requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
pillow
jinja2
openai  # or langchain / ollama
python-dotenv